fix: quick add not working for labels & assignee
